requirements.txt
amqp==5.0.9
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coverage==6.3.1
Deprecated==1.2.13
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
kombu==5.2.3
model-bakery==1.4.0
packaging==21.3
pep8==1.7.1
pluggy==1.0.0
prompt-toolkit==3.0.26
psycopg2==2.9.2
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.0
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-mock==3.7.0
python-decouple==3.6
pytz==2021.3
redis==4.1.2
requests==2.27.1
requests-mock==1.9.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2021.5
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.7.0
